// src/shared/api/models/getApiSalesListPrintTypeParams.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiSalesListPrintTypeSortBy } from './getApiSalesListPrintTypeSortBy';
import type { GetApiSalesListPrintTypeSalesType } from './getApiSalesListPrintTypeSalesType';

export type GetApiSalesListPrintTypeParams = {
    sortBy?: GetApiSalesListPrintTypeSortBy;
    fromDate?: string;
    toDate?: string;
    /**
     * Search by invoice number
     */
    keyword?: string;
    /**
     * Filter by sales type - SALES for regular sales, CREDIT for refunds/partial refunds
     */
    salesType?: GetApiSalesListPrintTypeSalesType;
};