// src/shared/api/models/getApiSalesListPrintTypeSortBy.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiSalesListPrintTypeSortBy =
    (typeof GetApiSalesListPrintTypeSortBy)[keyof typeof GetApiSalesListPrintTypeSortBy];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiSalesListPrintTypeSortBy = {
    ALL: 'ALL',
    CASH: 'CASH',
    CARD: 'CARD',
    GIFT_CARD: 'GIFT_CARD',
    CHEQUE: 'CHEQUE',
    ECOMMERCE: 'ECOMMERCE',
    OTHER: 'OTHER',
} as const;